backend/pyproject.toml
[project]
name = "kidsafe-alphabet-tutor-backend"
version = "0.1.0"
description = "Backend for KidSafe Alphabet Tutor"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "openai>=1.0.0",
    "anthropic",
    "gTTS",
    "SpeechRecognition",
    "pydub",
    "Pillow",
    "pytesseract",
    "fuzzywuzzy",
    "python-Levenshtein",
    "better-profanity",
    "pytest",
    "requests",
    "httpx",
    "python-dotenv",
    "websockets",
    "aiofiles",
    "sounddevice",
    "scipy",
    "livekit-agents[openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "opencv-python",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
